backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9

# Database & ORM
sqlalchemy>=2.0.31
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# RAG & AI
llama-index>=0.11.0
llama-index-vector-stores-qdrant>=0.3.0
llama-index-embeddings-ollama>=0.3.0
llama-index-llms-ollama>=0.3.0
pymupdf>=1.24.0

# Vector DB
qdrant-client>=1.9.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
structlog>=24.0.0
bleach>=6.1.0
python-magic>=0.4.27
